Gyro Sync feature (Loop synced to Gyro)

+++ src/main/drivers/accgyro_mpu6500.h
@@ -25,4 +25,4 @@ bool mpu6500AccDetect(acc_t *acc);
 bool mpu6500GyroDetect(gyro_t *gyro);
 
 void mpu6500AccInit(void);
-void mpu6500GyroInit(uint16_t lpf);
+void mpu6500GyroInit(uint8_t lpf);
